Changed classes to support creating rent report
