+++ src/main/java/com/rentCar/controller/TermController.java
@@ -7,6 +7,7 @@ import com.rentCar.service.TermService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -50,11 +51,43 @@ public class TermController {
                 advertisementDTO.setId(t.getAdvertisement().getId());
                 advertisementDTO.setCarBrand(t.getAdvertisement().getCar().getCarBrand().getName());
                 advertisementDTO.setName(t.getAdvertisement().getCar().getName());
+                advertisementDTO.setStartDate(t.getAdvertisement().getStartDate());
+                advertisementDTO.setEndDate(t.getAdvertisement().getEndDate());
 
                 termDTOS.add(termDTO);
             }
         }
 
+        return termDTOS;
+    }
+
+    @GetMapping(value = "/agent/{id}", produces = "application/json")
+    @PreAuthorize("hasRole('AGENT')")
+    public List<TermDTO> getAllRentedFromCurrentAgent(@PathVariable Long id) {
+        List<Term> rented = this.termService.getAllRentedFromCurrentAgent(id);
+        List<TermDTO> termDTOS = new ArrayList<>();
+
+        for (Term t : rented) {
+            TermDTO termDTO = new TermDTO();
+            termDTO.setId(t.getId());
+            termDTO.setStartDate(t.getStartDate());
+            termDTO.setEndDate(t.getEndDate());
+
+            AdvertisementDTO advertisementDTO = new AdvertisementDTO();
+            advertisementDTO.setId(t.getAdvertisement().getId());
+            advertisementDTO.setName(t.getAdvertisement().getCar().getName());
+            advertisementDTO.setCarBrand(t.getAdvertisement().getCar().getCarBrand().getName());
+            advertisementDTO.setModel(t.getAdvertisement().getCar().getCarModel().getName());
+            advertisementDTO.setStartDate(t.getAdvertisement().getStartDate());
+            advertisementDTO.setEndDate(t.getAdvertisement().getEndDate());
+
+            termDTO.setAdvertisement(advertisementDTO);
+
+            termDTOS.add(termDTO);
+
+        }
+
+        System.out.println(termDTOS.size());
         return termDTOS;
     }
